Slightly restructured front-end folders
