+++ frontend/src/assets/images/PulseNotesLogo.png
