+++ frontend/src/views/TranscriptionPage/TranscriptionPage.css
